pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jarvis"
version = "0.1.0"
description = "Always-On Ambient AI Copilot"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pyaudio>=0.2.13",
    "langfuse>=2.0.0",
]

[project.scripts]
jarvis = "jarvis.main:main"
